requirements.txt
streamlit
pandas
numpy
pillow
opencv-python
pytesseract
scikit-learn
reportlab
streamlit-drawable-canvas
